moving the createStylePostActionToast to dedicated file
